email confirm order, fix payment

+++ MyStore.Application/ModelView/SizeInStock.cs
@@ -2,7 +2,7 @@
 {
     public class SizeInStock
     {
-        public int SizeId { get; set; }
+        public long SizeId { get; set; }
         public string? SizeName { get; set; }
         public int InStock { get; set; }
     }
